Separated controller and view in there own package. Fixed error with how the wall was selected. Fixed Email.java so it was serializable allowing the program to run. Fixed Customer.java so it is serializable allowing the program to run.
