+++ 04 Implementation/app/src/main/java/com/example/newyorker/exceptions/CustomerAddressContainsSpecialCharacterException.java
@@ -1,4 +1,4 @@
-package com.example.newyorker.Exceptions;
+package com.example.newyorker.exceptions;
 
 public class CustomerAddressContainsSpecialCharacterException extends RuntimeException {
     public CustomerAddressContainsSpecialCharacterException(String message) {
