+++ 04 Implementation/app/src/main/java/com/example/newyorker/exceptions/CustomerEmailAddressWithoutAtSignException.java
@@ -1,4 +1,4 @@
-package com.example.newyorker.Exceptions;
+package com.example.newyorker.exceptions;
 
 public class CustomerEmailAddressWithoutAtSignException extends RuntimeException {
     public CustomerEmailAddressWithoutAtSignException(String message) {
