+++ 04 Implementation/app/src/main/java/com/example/newyorker/exceptions/CustomerPhoneNumberNotNumericException.java
@@ -1,4 +1,4 @@
-package com.example.newyorker.Exceptions;
+package com.example.newyorker.exceptions;
 
 public class CustomerPhoneNumberNotNumericException extends RuntimeException {
     public CustomerPhoneNumberNotNumericException(String message) {
