+++ 04 Implementation/app/src/main/java/com/example/newyorker/exceptions/CustomerZIPCodeNotFourDigitsException.java
@@ -1,4 +1,4 @@
-package com.example.newyorker.Exceptions;
+package com.example.newyorker.exceptions;
 
 public class CustomerZIPCodeNotFourDigitsException extends RuntimeException {
     public CustomerZIPCodeNotFourDigitsException(String message) {
