+++ 04 Implementation/app/src/main/java/com/example/newyorker/exceptions/CustomerZIPCodeNotNumericException.java
@@ -1,4 +1,4 @@
-package com.example.newyorker.Exceptions;
+package com.example.newyorker.exceptions;
 
 public class CustomerZIPCodeNotNumericException extends RuntimeException {
     public CustomerZIPCodeNotNumericException(String message) {
