+++ 04 Implementation/app/src/main/java/com/example/newyorker/exceptions/MissingCustomerEmailAddressException.java
@@ -1,4 +1,4 @@
-package com.example.newyorker.Exceptions;
+package com.example.newyorker.exceptions;
 
 public class MissingCustomerEmailAddressException extends RuntimeException {
     public MissingCustomerEmailAddressException(String message) {
