+++ 04 Implementation/app/src/main/java/com/example/newyorker/exceptions/MissingCustomerPhoneNumberException.java
@@ -1,4 +1,4 @@
-package com.example.newyorker.Exceptions;
+package com.example.newyorker.exceptions;
 
 public class MissingCustomerPhoneNumberException extends RuntimeException {
     public MissingCustomerPhoneNumberException(String message) {
