+++ 04 Implementation/app/src/main/java/com/example/newyorker/exceptions/MissingCustomerZIPCodeException.java
@@ -1,4 +1,4 @@
-package com.example.newyorker.Exceptions;
+package com.example.newyorker.exceptions;
 
 public class MissingCustomerZIPCodeException extends RuntimeException {
     public MissingCustomerZIPCodeException(String message) {
