+++ 04 Implementation/app/src/main/java/com/example/newyorker/exceptions/MissingWallHeightException.java
@@ -1,4 +1,4 @@
-package com.example.newyorker.Exceptions;
+package com.example.newyorker.exceptions;
 
 public class MissingWallHeightException extends RuntimeException {
     public MissingWallHeightException(String message) {
