+++ 04 Implementation/app/src/main/java/com/example/newyorker/exceptions/MissingWallWidthException.java
@@ -1,4 +1,4 @@
-package com.example.newyorker.Exceptions;
+package com.example.newyorker.exceptions;
 
 public class MissingWallWidthException extends RuntimeException {
     public MissingWallWidthException(String message) {
