+++ 04 Implementation/app/src/main/java/com/example/newyorker/exceptions/WallHeightNotNumericException.java
@@ -1,4 +1,4 @@
-package com.example.newyorker.Exceptions;
+package com.example.newyorker.exceptions;
 
 public class WallHeightNotNumericException extends RuntimeException {
     public WallHeightNotNumericException(String message) {
