+++ 04 Implementation/app/src/main/java/com/example/newyorker/exceptions/WallHeightTooHighException.java
@@ -1,4 +1,4 @@
-package com.example.newyorker.Exceptions;
+package com.example.newyorker.exceptions;
 
 public class WallHeightTooHighException extends RuntimeException {
     public WallHeightTooHighException(String message) {
