+++ 04 Implementation/app/src/main/java/com/example/newyorker/exceptions/WallWidthTooSmallException.java
@@ -1,4 +1,4 @@
-package com.example.newyorker.Exceptions;
+package com.example.newyorker.exceptions;
 
 public class WallWidthTooSmallException extends RuntimeException {
     public WallWidthTooSmallException(String message) {
