+++ 04 Implementation/app/src/main/java/com/example/newyorker/model/Wall.java
@@ -109,6 +109,8 @@ public class Wall  implements Serializable{
     public void removeAllObservers(){
         COLLECTIONOFWALLOBSERVERS.removeAll(COLLECTIONOFWALLOBSERVERS);
     }
+
+
     //</editor-fold>
 
 }
