+++ 04 Implementation/app/src/main/java/com/example/newyorker/view/PreviewOrderActivity.java
@@ -1,4 +1,4 @@
-package com.example.newyorker;
+package com.example.newyorker.view;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -7,10 +7,9 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
 
-import com.example.newyorker.model.NYBuilderController;
+import com.example.newyorker.R;
+import com.example.newyorker.controller.NYBuilderController;
 import com.example.newyorker.model.Observer;
-import com.example.newyorker.model.Specifications;
-import com.example.newyorker.model.Wall;
 
 public class PreviewOrderActivity extends AppCompatActivity {
