+++ 04 Implementation/app/src/main/res/layout-v21/activity_customize_order.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CustomizeOrderActivity">
+    tools:context=".view.CustomizeOrderActivity">
 
     <TextView
         android:id="@+id/textview_price_customize_activity"
