+++ 04 Implementation/app/src/main/res/layout-v21/activity_preview_order.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PreviewOrderActivity">
+    tools:context=".view.PreviewOrderActivity">
 
     <Button
         android:id="@+id/button_wall3"
