+++ 04 Implementation/app/src/main/res/layout/activity__send__request.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SendRequestActivity">
+    tools:context=".view.SendRequestActivity">
 
     <ImageView
         android:id="@+id/imageView2"
