+++ 04 Implementation/app/src/main/res/layout/activity_customize_order.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CustomizeOrderActivity">
+    tools:context=".view.CustomizeOrderActivity">
 
     <Spinner
         android:id="@+id/spinner_doors"
