+++ 04 Implementation/build.gradle
@@ -14,9 +14,6 @@ buildscript {
 
 allprojects {
     repositories {
-        maven{
-            url"https://jitpack.io"
-        }
         google()
         jcenter()
     }
